Created objects and updated game's logic
Created an object for each card in the cards array.
Each objects stores the card's name (property rank),
the suit, and a path to where the image for that
card is stored (property cardImage). Then updated
the code from the last assignment, in order to access
the cards values from their specific cards object.