Finished game's logic using DOM
When the user clicks on a card, the image
with the face of that card is displayed.
If the user selectes two cards, an alert
pops up letting the user know if the two
cards match. Also added a reset game button
that reloads the page to the HTML code.